[Feature 1][Refactor] Separate AgeInput, AgeOutput components
