+++ src/AgeInput.jsx
@@ -1,4 +1,5 @@
-import React, { useState, useEffect } from "react";
+import React, { useState } from "react";
+import AgeOutput from "./AgeOutput.jsx";
 
 const AgeInput = () => {
   const currentYear = new Date().getFullYear();
@@ -8,43 +9,49 @@ const AgeInput = () => {
   const [selectedYear, setSelectedYear] = useState(currentYear);
   const [selectedMonth, setSelectedMonth] = useState(1);
   const [selectedDay, setSelectedDay] = useState(1);
-  const [koreanAge, setKoreanAge] = useState(null); // 만 나이
-  const [internationalAge, setInternationalAge] = useState(null); // 연 나이
-  const [formerKoreanAge, setFormerKoreanAge] = useState(null); // 세는 나이
+  // const [koreanAge, setKoreanAge] = useState(null); // 만 나이
+  // const [internationalAge, setInternationalAge] = useState(null); // 연 나이
+  // const [formerKoreanAge, setFormerKoreanAge] = useState(null); // 세는 나이
 
-  useEffect(() => {
-    calculateKoreanAge();
-  }, [selectedMonth, selectedYear, selectedDay]);
+  // useEffect(() => {
+  //   calculateKoreanAge();
+  // }, [selectedMonth, selectedYear, selectedDay]);
 
-  const calculateKoreanAge = () => {
-    const today = new Date();
-    const birthDate = new Date(selectedYear, selectedMonth - 1, selectedDay);
+  // const calculateKoreanAge = () => {
+  //   const today = new Date();
+  //   const birthDate = new Date(selectedYear, selectedMonth - 1, selectedDay);
 
-    let age = today.getFullYear() - birthDate.getFullYear();
-    console.log(age);
+  //   let age = today.getFullYear() - birthDate.getFullYear();
+  //   console.log(age);
 
-    setInternationalAge(age);
-    setFormerKoreanAge(age + 1);
-    // 만 나이 계산
-    if (
-      today.getMonth() < birthDate.getMonth() ||
-      (today.getMonth() === birthDate.getMonth() && today.getDate() < birthDate.getDate())
-    ) {
-      age--;
-    }
-    setKoreanAge(age);
-  };
+  //   setInternationalAge(age);
+  //   setFormerKoreanAge(age + 1);
+  //   // 만 나이 계산
+  //   if (
+  //     today.getMonth() < birthDate.getMonth() ||
+  //     (today.getMonth() === birthDate.getMonth() && today.getDate() < birthDate.getDate())
+  //   ) {
+  //     age--;
+  //   }
+  //   setKoreanAge(age);
+  // };
 
   const handleYearChange = (e) => {
     setSelectedYear(parseInt(e.target.value));
+    // const targetValue = parseInt(e.target.value);
+    // selectedYear !== targetValue ? setSelectedYear(parseInt(e.target.value)) : 0;
   };
 
   const handleMonthChange = (e) => {
     setSelectedMonth(parseInt(e.target.value));
+    // const targetValue = parseInt(e.target.value);
+    // selectedMonth !== targetValue ? setSelectedMonth(parseInt(e.target.value)) : 0;
   };
 
   const handleDayChange = (e) => {
     setSelectedDay(parseInt(e.target.value));
+    // const targetValue = parseInt(e.target.value);
+    // selectedDay !== targetValue ? setSelectedDay(parseInt(e.target.value)) : 0;
   };
 
   return (
@@ -78,13 +85,14 @@ const AgeInput = () => {
         )}
       </select>
 
-      {koreanAge !== null && (
+      {/* {koreanAge !== null && (
         <div>
           <p>만 나이 Korean Age: {koreanAge - 1 < 0 ? 0 : koreanAge}</p>
           <p>연 나이 Korean Age: {internationalAge}</p>
           <p>세는 나이 Korean Age: {formerKoreanAge}</p>
         </div>
-      )}
+      )} */}
+      <AgeOutput data={{ selectedYear: selectedYear, selectedDay: selectedDay, selectedMonth: selectedMonth }} />
     </div>
   );
 };
