+++ src/App.jsx
@@ -1,12 +1,12 @@
 import "./App.css";
 import AgeInput from "./AgeInput.jsx";
 
-function App() {
+const App = () => {
   return (
     <>
       <AgeInput />
     </>
   );
-}
+};
 
 export default App;
